Move buffered and borrowing parts of reader to separate files
This commit only moves code without significant changes (the only changes is:
- corrected imports
- add imports to the doc comments which have become inaccessible
)
